Added Configurations, Exceptions, Midllewares files, then corrected P…

+++ src/SpeakHub.Service/Interfaces/Common/IIdentityService.cs
@@ -2,7 +2,7 @@
 {
     public interface IIdentityService
     {
-        public long? Id { get; }
+        public int? Id { get; }
 
         public string FirstName { get; }
